Added MAFFT

+++ ViralPrimerDesign.py
@@ -10,14 +10,16 @@ from math import log2
 from os import mkdir
 from os.path import abspath, expanduser, isdir, isfile
 from seaborn import relplot
+from subprocess import call, check_output, STDOUT
 from sys import argv, stdin, stderr
 import argparse
 import matplotlib.pyplot as plt
 
 # constants
 DEFAULT_BUFSIZE = 1048576 # 1 MB
-NUCS_SORT = ['A', 'C', 'G', 'T', '-']; NUCS = set(NUCS_SORT)
+MIN_MAFFT_VERSION = 7.467
 NUC_COLORS = {'A':'red', 'C':'blue', 'G':'purple', 'T':'yellow', '-':'black'}
+NUCS_SORT = ['A', 'C', 'G', 'T', '-']; NUCS = set(NUCS_SORT)
 NUM_SEQS_PROGRESS = 500
 
 # helper class for logging
@@ -49,8 +51,10 @@ def parse_args():
     # use argparse to parse user arguments
     parser = argparse.ArgumentParser(description=__doc__, formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument('-s', '--sequences', required=False, type=str, default='stdin', help="Input Sequences (FASTA format)")
+    parser.add_argument('-r', '--reference', required=True, type=str, help="Reference Genome (FASTA format)")
     parser.add_argument('-o', '--outdir', required=True, type=str, help="Output Directory")
     parser.add_argument('--skip_alignment', action="store_true", help="Skip Alignment (if input FASTA is already aligned)")
+    parser.add_argument('-t', '--threads', required=False, type=int, default=1, help="Number of Threads (for MSA)")
     parser.add_argument('-q', '--quiet', action="store_true", help="Quiet (hide verbose messages)")
     parser.add_argument('-v', '--version', action="store_true", help="Show Version Number")
     args = parser.parse_args()
@@ -58,15 +62,45 @@ def parse_args():
     # check user arguments for validity
     if args.sequences != 'stdin' and not isfile(args.sequences):
         raise ValueError("File not found: %s" % args.sequences)
+    if not isfile(args.reference):
+        raise ValueError("File not found: %s" % args.reference)
     args.outdir = abspath(expanduser(args.outdir))
     if isdir(args.outdir) or isfile(args.outdir):
         raise ValueError("Output exists: %s" % args.outdir)
+    if args.threads < 1:
+        raise ValueError("Number of threads must be positive integer: %s" % args.threads)
     return args
+
+# align using MAFFT
+def align_mafft(in_fn, ref_fn, out_fn, threads=1, logger=None):
+    try:
+        mafft_version = check_output(['mafft', '--version'], stderr=STDOUT).decode()
+    except:
+        raise RuntimeError("Unable to execute 'mafft'. Are you sure it's in your PATH?")
+    mafft_version = float(mafft_version.split()[0].lstrip('v'))
+    if mafft_version < MIN_MAFFT_VERSION:
+        raise RuntimeError("Must use MAFFT v%s or higher, but detected v%s" % (MIN_MAFFT_VERSION, mafft_version))
+    for fn in [in_fn, ref_fn]:
+        if not isfile(fn):
+            raise ValueError("File not found: %s" % fn)
+    if isfile(out_fn) or isdir(out_fn):
+        raise ValueError("Output exists: %s" % fn)
+    command = ['mafft', '--thread', str(threads), '--6merpair', '--addfragments', in_fn, ref_fn]
+    if logger is not None:
+        logger.write("Aligning sequences from: %s\n" % in_fn)
+        logger.write("Reference genome: %s\n" % ref_fn)
+        logger.write("MAFFT command: %s\n" % ' '.join(command))
+    o = open(out_fn, 'w'); e = open('%s/mafft.log' % '/'.join(out_fn.split('/')[:-1]), 'w')
+    call(command, stdout=o, stderr=e); o.close(); e.close()
+    if logger is not None:
+        logger.write("Multiple sequence alignment written to: %s\n" % out_fn)
 
 # iterate over sequences of FASTA
 def iter_fasta(in_fn):
     if in_fn == 'stdin':
         in_f = stdin
+    elif not isfile(in_fn):
+        raise ValueError("File not found: %s" % in_fn)
     elif in_fn.lower().endswith('.gz'):
         in_f = gopen(in_fn, 'rt')
     else:
@@ -104,8 +138,8 @@ def count_bases(in_fn, logger=None):
 
 # write base counts
 def write_counts(counts, out_fn, delim='\t', logger=None):
-    if isfile(out_fn):
-        raise ValueError("File exists: %s" % out_fn)
+    if isfile(out_fn) or isdir(out_fn):
+        raise ValueError("Output exists: %s" % out_fn)
     if logger is not None:
         logger.write("Writing base counts to file...\n")
     out_f = open(out_fn, 'w'); out_f.write('Position (0-indexed)%s%s\n' % (delim, delim.join(NUCS_SORT)))
@@ -140,8 +174,8 @@ def compute_entropies(counts, logger=None):
 
 # write Shannon entropies
 def write_entropies(ents, out_fn, delim='\t', logger=None):
-    if isfile(out_fn):
-        raise ValueError("File exists: %s" % out_fn)
+    if isfile(out_fn) or isdir(out_fn):
+        raise ValueError("Output exists: %s" % out_fn)
     if logger is not None:
         logger.write("Writing entropies to file...\n")
     out_f = open(out_fn, 'w'); out_f.write('Position (0-indexed)%sEntropy\n' % delim)
@@ -153,8 +187,8 @@ def write_entropies(ents, out_fn, delim='\t', logger=None):
 
 # plot entropies
 def plot_entropies(ents, out_fn, logger=None):
-    if isfile(out_fn):
-        raise ValueError("File exists: %s" % out_fn)
+    if isfile(out_fn) or isdir(out_fn):
+        raise ValueError("Output exists: %s" % out_fn)
     if logger is not None:
         logger.write("Plotting entropies...\n")
     x = [i for i, e in enumerate(ents) if e is not None]
@@ -172,7 +206,8 @@ if __name__ == "__main__":
     logger.write("=== ViralPrimerDesign v%s ===\n" % VERSION)
     logger.write("Command: %s\n" % ' '.join(argv))
     if not args.skip_alignment:
-        args.sequences = align_mafft(args.sequences, '%s/sequences.aln' % args.outdir, logger=logger)
+        orig_sequences = args.sequences; args.sequences = '%s/sequences.aln' % args.outdir
+        align_mafft(orig_sequences, args.reference, args.sequences, threads=args.threads, logger=logger)
     counts = count_bases(args.sequences, logger=logger)
     write_counts(counts, '%s/counts.tsv' % args.outdir, logger=logger)
     consensus = compute_consensus(counts, '%s/consensus.txt' % args.outdir, logger=logger)
